+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added splash screen and multi-language support to the desktop application.
 - Created the about dialog to show the application information on the desktop.
 - Created the new project dialog to allow the user to create a new project.
+- Created the open project dialog to allow the user to open an existing project.
 
 ### Fix